gh-657 Add StatisticsSearchService and TestFlowable
