+++ stroom-statistics-server/src/main/java/stroom/statistics/server/common/search/StatisticsSearchService.java
@@ -1,0 +1,13 @@
+package stroom.statistics.server.common.search;
+
+import io.reactivex.Flowable;
+import stroom.dashboard.expression.FieldIndexMap;
+import stroom.statistics.common.FindEventCriteria;
+import stroom.statistics.shared.StatisticStoreEntity;
+
+public interface StatisticsSearchService {
+
+    Flowable<String[]> search(final StatisticStoreEntity statisticStoreEntity,
+                              final FindEventCriteria criteria,
+                              final FieldIndexMap fieldIndexMap);
+}
